handel err

+++ view/music/list.php
@@ -3,12 +3,12 @@
 ?>
 <?php
 
-$id="";
+$id = "";
 ?>
 
 <?php ob_start()   ?>
 <br>
-<center><a href="http://localhost/pdomvc/index.php/music" style="font-size: medium; color:grey;cursor: pointer;">Home</a></center>
+<a href="http://localhost/pdomvc/index.php/music" style="font-size: medium; color:grey;cursor: pointer;">Return</a>
 <br>
 <br>
 <center>
@@ -16,15 +16,15 @@ $id="";
 </center>
 <br>
 
-<form action="http://localhost/pdomvc/index.php/music/search?id=<?+ $_GET['id'] ?>" >
+<form action="http://localhost/pdomvc/index.php/music/search?id=<? +$_GET['id'] ?>">
     <div style="margin-bottom: 20px;">
-        <input type="text" name="id" value="<?= $id ?>"   required/>
-       
+        <input type="text" name="id" value="<?= $id ?>" required />
+
         <input type="submit" />
     </div>
-   </form>
-  
-   <?php echo $id?> 
+</form>
+
+<?php echo $id ?>
 
 
 <div class="table-responsive">
@@ -40,20 +40,27 @@ $id="";
             <th>Edit</th>
             <th>Delete</th>
         </tr>
-        <?php foreach ($music as $row) : ?>
 
-            <tr>
-                <td><?= $row['id'] ?></td>
-                <td><?= $row['nama'] ?></td>
-                <td><?= $row['judul'] ?></td>
-                <td><?= $row['album'] ?></td>
-                <td><?= $row['tahun'] ?></td>
-                <td><a href="http://localhost/pdomvc/index.php/music/detail?id=<?= $row['id'] ?>" class="btn btn-success btn-xs"> Detail</a></td>
-                <td><a href="http://localhost/pdomvc/index.php/music/edit?id=<?= $row['id'] ?>" class="btn btn-warning btn-xs"><span class="glyphicon glyphicon-edit"></span> Edit</a></td>
-                <td><a href="http://localhost/pdomvc/index.php/music/delete?id=<?= $row['id'] ?>" onclick="return confirm('Anda yakin akan menghapus data ini?')" class="btn btn-danger btn-xs"> <span class="glyphicon glyphicon-trash"></span> Delete</a></td>
-            </tr>
-        <?php endforeach;
+        <?php if ($music[0]) : foreach ($music as $row) : ?>
 
+                <tr>
+                    <td><?= $row['id'] ?></td>
+                    <td><?= $row['nama'] ?></td>
+                    <td><?= $row['judul'] ?></td>
+                    <td><?= $row['album'] ?></td>
+                    <td><?= $row['tahun'] ?></td>
+                    <td><a href="http://localhost/pdomvc/index.php/music/detail?id=<?= $row['id'] ?>" class="btn btn-success btn-xs"> Detail</a></td>
+                    <td><a href="http://localhost/pdomvc/index.php/music/edit?id=<?= $row['id'] ?>" class="btn btn-warning btn-xs"><span class="glyphicon glyphicon-edit"></span> Edit</a></td>
+                    <td><a href="http://localhost/pdomvc/index.php/music/delete?id=<?= $row['id'] ?>" onclick="return confirm('Anda yakin akan menghapus data ini?')" class="btn btn-danger btn-xs"> <span class="glyphicon glyphicon-trash"></span> Delete</a></td>
+                </tr>
+            <?php endforeach;
+        else :
+            ?>
+            <div class="alert alert-danger" role="alert">
+                wlh akhoya makhbito alik hi 3awd 9alab
+            </div>
+        <?php
+        endif;
         ?>
     </table>
 </div>
